# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: FabMind Agent API
  version: 1.0.0
  description: On-premise evidence-first troubleshooting API for Load Port / FOUP Clamp equipment.
servers:
  - url: http://localhost:8000
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/login:
    post:
      summary: Demo login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/auth/me:
    get:
      summary: Current authenticated user
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }
    LoginResponse:
      type: object
      required: [access_token, user]
      properties:
        access_token: { type: string }
        token_type: { type: string, enum: [bearer] }
        user: { $ref: '#/components/schemas/User' }
    User:
      type: object
      required: [id, username, display_name, role, tenant_id]
      properties:
        id: { type: string }
        username: { type: string }
        display_name: { type: string }
        role: { type: string, enum: [FIELD_ENGINEER, SENIOR_ENGINEER, ADMIN] }
        tenant_id: { type: string }
...
    ErrorResponse:
      type: object
      required: [error_code, message]
      properties:
        error_code: { type: string }
        message: { type: string }
        details: { type: object }
        request_id: { type: string }